tests/requirements.txt
pytest
requests
pyjwt
selenium